backend/requirements.txt
api==0.0.7
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
coverage==7.3.0
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.4
django-colorfield==0.9.0
django-debug-toolbar==3.2.4
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
html5lib==1.1
idna==3.4
lxml==4.9.3
nose==1.3.7
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==10.0.0
pycparser==2.21
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==2.8.0
pypdf==3.15.5
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.4
webencodings==0.5.1
